% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylogeny_tree.R
\name{add_leaf_list}
\alias{add_leaf_list}
\title{Phylogenetic tree leaf}
\usage{
add_leaf_list(leaf, connexion_list, timepoints, d, selector_position)
}
\arguments{
\item{leaf}{A vector of cellularities (ranging from 0 to 1)}

\item{connexion_list}{List containing 1. An interaction matrix concatenated with the cellularity of each cluster (one line per cluster)}

\item{timepoints}{A numeric vector giving the spatial and/or temporal distribution of the samples}

\item{d}{The initial number of clusters}

\item{selector_position}{The row of the studied leaf in the data frame.}
}
\description{
Adds a leaf to an already built tree. Output is a list of all possibilities.
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}
