% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{BIC_criterion}
\alias{BIC_criterion}
\title{Bayesian Information Criterion}
\usage{
BIC_criterion(EM_out_list, model.selection)
}
\arguments{
\item{EM_out_list}{list of outputs from EM.algo or FullEM}

\item{model.selection}{The function to minimize for the model selection: can be "AIC", "BIC", or numeric. In numeric, the function}
}
\description{
Computes BIC from a list of outputs of EM algorithm, then returns the position with minimal BIC
}
\keyword{EM}
\keyword{clustering}
\keyword{number}

