% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{BIC_criterion_FLASH}
\alias{BIC_criterion_FLASH}
\title{Compute criterion FLASH}
\usage{
BIC_criterion_FLASH(Obj, Mut_num, k, model.selection, s)
}
\arguments{
\item{Obj}{Numeric vector with objective function values}

\item{Mut_num}{Number of mutations to cluster}

\item{k}{the number of clusters (in the same order as Obj)}

\item{model.selection}{The function to minimize for the model selection: can be "AIC", "BIC", or numeric. In numeric, the function}

\item{s}{Number of samples}
}
\description{
Computes BIC from a list of outputs of EM algorithm, then returns the position with minimal BIC
}
\keyword{EM}
\keyword{clustering}
\keyword{number}
