% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{Cellular_preclustering}
\alias{Cellular_preclustering}
\title{Preclustering method}
\usage{
Cellular_preclustering(Schrod_cells)
}
\arguments{
\item{Schrod_cells}{The classic output from Schrodinger function}
}
\value{
returns a list with:
\describe{
 \item{similarityMatrix}{ The matrix of probabilities}
 \item{distance}{The dissimilarity matrix}
 \item{tree}{The tree obtained by hierachical clustering of the dissimilarity matrix using "ward.D2" method}
 }
}
\description{
This method clusters mutations based on the probability that they are from the same 
distribution.
It first computes the zscore associated with a normalized number of alternative reads and depth.
The "normalized" number of reads is the number of alternative reads expected if the mutation was at a single 
copy in a diploid genome.
}
