% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_generation.R
\name{check_split}
\alias{check_split}
\title{Check}
\usage{
check_split(leaf)
}
\arguments{
\item{leaf}{Numeric vector}
}
\description{
Check if node can be split further, i.e. cellularity > 10% in at least 1 sample
}

