% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{Cluster_drop}
\alias{Cluster_drop}
\title{Find clusters to fuse}
\usage{
Cluster_drop(centers, epsilon)
}
\arguments{
\item{centers}{list or vector with centers of clusters}

\item{epsilon}{maximal distance in each sample between two cluster centers to fuse}
}
\value{
Returns numbers of clusters to drop
}
\description{
Find clusters to fuse
}

