% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantumCat.R
\name{Compute_NMI}
\alias{Compute_NMI}
\title{Normalized Mutual Information}
\usage{
Compute_NMI(QC_out)
}
\arguments{
\item{QC_out}{output from QuantumClone clustering}
}
\description{
Compute the normalized mutual information to assess clustering quality
}
\examples{
Compute_NMI(QC_output)
}

