% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{Compute_objective}
\alias{Compute_objective}
\title{Compute value of objective function}
\usage{
Compute_objective(tree, nclus, Schrod, conta)
}
\arguments{
\item{tree}{Tree from hierarchical clustering}

\item{nclus}{Number of clusters used for cutting (numeric of length 1)}

\item{Schrod}{Output from Schrodinger cellularities}

\item{conta}{Numeric value of contamination fraction in each sample}
}
\description{
Compute the value of clustering based on same principles as QuantumClone EM
}
