% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EM.R
\name{create_priors}
\alias{create_priors}
\title{Clonal fraction prior creation}
\usage{
create_priors(nclust, nsample, prior = NULL)
}
\arguments{
\item{nclust}{Number of clones to look for.}

\item{nsample}{Number of samples}

\item{prior}{Possible priors known (the position of each element in a list corresponds to 1 clone)}
}
\description{
Semi-random generation of clonal priors
}
\keyword{EM}

