% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM.algo}
\alias{EM.algo}
\title{Expectation Maximization algorithm}
\usage{
EM.algo(Schrod, nclust = NULL, prior_center = NULL, prior_weight = NULL,
  contamination, epsilon = 10^(-2), optim = "default", integrate)
}
\arguments{
\item{Schrod}{A list of dataframes (one for each sample), generated by the Patient_schrodinger_cellularities() function.}

\item{nclust}{Number of clones to look for (mandatory if prior_center or prior_weight are null)}

\item{prior_center}{Clone coordinates (from another analysis) to be used}

\item{prior_weight}{Prior on the fraction of mutation in each clone}

\item{contamination}{Numeric vector with the fraction of normal cells contaminating the sample}

\item{epsilon}{Stopping condition for the algorithm: what is the minimal tolerated difference of position or weighted between two steps}

\item{optim}{use L-BFS-G optimization from R ("default"), or from optimx ("optimx")}

\item{integrate}{Should we perform integration of density over ~epsilon sized interval}

\item{epsilon}{Stop value: maximal admitted value of the difference in cluster position and weights 
between two optimization steps. If NULL, will take 1/(median depth). Also used for integration size.}
}
\description{
Optimization of clone positions and proportion of mutations in each clone.
}
\keyword{EM}

