% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{eval.fik.m}
\alias{eval.fik.m}
\title{Eval probability for M step
Computes the log directly as log density is faster to compute}
\usage{
eval.fik.m(Schrod, centers, weights, alpha, adj.factor, epsilon, integrate)
}
\arguments{
\item{Schrod}{The shcrodinger list of matrices}

\item{centers}{centers of the clusters}

\item{weights}{weight of each cluster}

\item{alpha}{weight of each status (number of copies for a mutation)}

\item{adj.factor}{The adjusting factor, taking into account contamination, copy number, number of copies}

\item{epsilon}{Stop value: maximal admitted value of the difference in cluster position and weights 
between two optimization steps. If NULL, will take 1/(median depth). Also used for integration size.}

\item{integrate}{Should QuantumClone integrate probabilities over epsilon interval?}
}
\description{
Eval probability for M step
Computes the log directly as log density is faster to compute
}

