% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{evolution_plot}
\alias{evolution_plot}
\title{Evolution plot}
\usage{
evolution_plot(QC_out, Sample_names = NULL)
}
\arguments{
\item{QC_out}{: Output from One_step_clustering}

\item{Sample_names}{: character vector of the names of each sample (in the same order as the data)}
}
\description{
Plots evolution in time of clones
}
\examples{
require(ggplot2)
}

