% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{filter_on_fik}
\alias{filter_on_fik}
\title{Data filter}
\usage{
filter_on_fik(Schrod, fik)
}
\arguments{
\item{Schrod}{A list of dataframes (one for each sample), generated by the Patient_schrodinger_cellularities.}

\item{fik}{matrix of probability of each possibility to belong to a clone}
}
\description{
Keep one possibility per position and ajust weight accordingly
}
\keyword{filter}

