% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{FLASH_main}
\alias{FLASH_main}
\title{Flash core}
\usage{
FLASH_main(Schrod_cells, model.selection, conta, Nclus, tree = NULL,
  dissimMatrix = NULL)
}
\arguments{
\item{Schrod_cells}{Output from Schrodinger cellularities}

\item{model.selection}{One of "tree", "AIC", "BIC" or numeric. "tree" will use "ccc","ch" and "gap" methods from NbClust 
to determine the number of clusters. "BIC","AIC" or numeric values will use methods from QuantumClone.}

\item{conta}{vector with contamination fraction in each sample}

\item{Nclus}{vector with the number of clusters to test (alternatively only min and max values)}

\item{tree}{Hierarchical tree from hclust}

\item{dissimMatrix}{Dissimilarity matrix, required if model selection is "tree"}
}
\description{
Returns number of clusters based on model selection
}
