% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descent.R
\name{grzero}
\alias{grzero}
\title{Gradient 0}
\usage{
grzero(fik, adj.factor, Alt, Depth)
}
\arguments{
\item{fik}{matrix with probability of each possibility to belong to clone k}

\item{adj.factor}{matrix with coefficient making transition between cellularity and frequency}

\item{Alt}{matrix with samples in columns and number of alternative reads in rows}

\item{Depth}{matrix with samples in coluns and depth of coverage in rows}
}
\description{
Return center values for max if adj.factor has a single value for all variants/possibilities in each samples
}

