% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Input_Example}
\alias{Input_Example}
\title{Example generated by QuantumCat}
\format{A list of dataframes
\describe{
 \item{SampleName}{First column containing the name of the sample}
 \item{Chr}{The chromosome either 1 2 ... X Y or chr1 chr2 ... chrY}
 \item{Start}{The genomic position of the variant}
 \item{Alt}{Number of reads supporting variant allele}
 \item{Depth}{Total number of reads (reference + alternative allele) at position}
}}
\usage{
Input_Example
}
\description{
A dataset generated by:
Input_Example<-QuantumCat(number_of_clones = 4,
                         number_of_mutations = 100,
                         ploidy = "AB",depth = 150,
                         number_of_samples = 2,
                         contamination = c(0,0))
}
\keyword{datasets}

