% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumClone_v3.R
\name{is_included}
\alias{is_included}
\title{Group theory}
\usage{
is_included(Clone1, Clone2)
}
\arguments{
\item{Clone1}{Numeric vector, representing the cellularity of Clone1 in different samples}

\item{Clone2}{Numeric vector, representing the cellularity of Clone2 in different samples}
}
\description{
Clone2 is included in Clone1 if all values of Clone2 are lower or equal to the ones in Clone1 at the same position. Returns TRUE is Clone2 is included in Clone1.
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}

