% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descent.R
\name{list_prod}
\alias{list_prod}
\title{List product}
\usage{
list_prod(L, col = NULL)
}
\arguments{
\item{L}{list used}

\item{col}{If it is a list of matrices, and only one column should be used, name of the column.}
}
\description{
Returns the product of all elements in a list, e.g. a vector if the elements of the list are vectors, etc.
}
\keyword{2)))}
\keyword{2),matrix(1:4,nrow}
\keyword{=}
\keyword{List}
\keyword{handling}
\keyword{list_prod(list(matrix(1:4,nrow}

