% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumClone_v3.R
\name{longueur}
\alias{longueur}
\title{Length}
\usage{
longueur(matrix, n)
}
\arguments{
\item{matrix}{The interaction matrix of the tree (1 on the i-th row j-th column means "clone j is the progeny of clone i")}

\item{n}{Index of the clone in the matrix}
}
\description{
Computes the length from the clone on the n-th row of the matrix, to the most ancestral clone
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}

