% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{MajorityVote}
\alias{MajorityVote}
\title{Majority vote}
\usage{
MajorityVote(index)
}
\arguments{
\item{index}{vector with number of clusters selected by indices}
}
\value{
Numeric value of the number of clusters to chose
}
\description{
Extract majority vote from multiple indices
}
