% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumClone_v3.R
\name{manual_plot_trees}
\alias{manual_plot_trees}
\title{Plot tree}
\usage{
manual_plot_trees(connexion_list, d, cex = 0.8, p)
}
\arguments{
\item{connexion_list}{Data frame of the concatenation of the interaction matrix and the cellularity of each clone at different time points.}

\item{d}{Initial number of clones}

\item{cex}{Coefficient of expansion for the texts in phylogenetic tree plots. Default is 0.8}

\item{p}{Probability of a tree}
}
\description{
Creates a visual output for the phylogeny created by Tree_generation()
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}

