% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumClone_v3.R
\name{multiplot_trees}
\alias{multiplot_trees}
\title{Plots multiple trees}
\usage{
multiplot_trees(result_list, d, cex = 0.8)
}
\arguments{
\item{result_list}{List of lists (tree generated and the probability associated with each tree)}

\item{d}{Initial number of clones}

\item{cex}{Coefficient of expansion for the texts in phylogenetic tree plots. Default is 0.8}
}
\description{
Plots all trees created by the function Tree_generation. The red line means that mutations occured.
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}

