% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumClone_v3.R
\name{One_D_plot}
\alias{One_D_plot}
\title{Plots}
\usage{
One_D_plot(EM_out, contamination)
}
\arguments{
\item{EM_out}{output from the EM algorithm}

\item{contamination}{Numeric vector giving the proportion of normal cells in each samples}
}
\description{
Creates density plot when only one sample is given
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}

