% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descent.R
\name{peak}
\alias{peak}
\title{Peak}
\usage{
peak(x, y, prob, epsilon)
}
\arguments{
\item{x}{Number of alternative reads}

\item{y}{Number of draws}

\item{prob}{probability to draw}

\item{epsilon}{coefficient to normalize interval size}
}
\description{
Integrate binomial density over 2epsilon*depth interval
}

