% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_generation.R
\name{phylo_tree_generation}
\alias{phylo_tree_generation}
\title{Data generation}
\usage{
phylo_tree_generation(number_of_clones, number_of_samples)
}
\arguments{
\item{number_of_clones}{The wanted number of observable clones (meaning bearing at least 1 mutation)}

\item{number_of_samples}{The number of samples on which the data should be simulated}
}
\description{
Creates a phylogenetic tree on simple assumptions:
1. There is a common ancestor to all clones
2. Each clone creates a partition of the space. One node will carry mutations, the other will not. 
A node that is not mutated can be partitioned.
3. Nodes that have less than 2% difference in all samples to another existing clone will be discarded or
less than 1% cellularity in all samples.
4. Nodes that have less than 10% cellularity in ALL samples cannot be split further.
WARNING: Tree_generation recreates a tree from data while phylo_tree_generation randomly creates a phylogeny
}
\keyword{Data}
\keyword{creation}
\keyword{phylogeny}
