% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumClone_v3.R
\name{plot_cell_from_Return_out}
\alias{plot_cell_from_Return_out}
\title{Plot cellularity}
\usage{
plot_cell_from_Return_out(lis, Sample_names, output_dir = NULL)
}
\arguments{
\item{lis}{Output from Return_one_cell_by_mut, list of cellularities (one list-element per sample)}

\item{Sample_names}{Name of the samples.}

\item{output_dir}{Directory in which to save plots}
}
\description{
2D plot of cellularity based on the output of the EM
}
\keyword{Clonal}
\keyword{inference}
\keyword{plot}

