% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_QC_out}
\alias{plot_QC_out}
\title{Plot QC_output}
\usage{
plot_QC_out(QClone_Output, Sample_names = NULL, simulated = FALSE,
  sample_selected = 1:2)
}
\arguments{
\item{QClone_Output}{Output from QuantumClone algorithm}

\item{Sample_names}{: character vector of the names of each sample (in the same order as the data)}

\item{simulated}{Was the data generated by QuantumCat?}

\item{sample_selected}{: number of the sample to be considered for plot (can be 1 or 2 samples)}
}
\description{
This function was implemented to re-plot easily the diagrams of clonality for changes/enhancement.
Returns a ggplot object
Uses ggplot2 package
}
\examples{
require(ggplot2)
message("Using preclustered data:")
QC_out<-QuantumClone::QC_output
plot_QC_out(QC_out,Sample_names = c("Diagnosis","Relapse"))

}
\keyword{Plot}

