% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{ProbDistMatrix}
\alias{ProbDistMatrix}
\title{Distance}
\usage{
ProbDistMatrix(Schrod_cells)
}
\arguments{
\item{Schrod_cells}{The classic output from Schrodinger function, with the normalized Alt}
}
\value{
returns a square numeric matrix
}
\description{
Creates a matrix of distance between two points based on the p-value to be from the same distribution
}

