% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QC_output}
\alias{QC_output}
\title{Example of output generated by clustering of Input_Example}
\format{list of lists
\describe{
 \item{EMOutput}{Results of Expectation maximization:
   *fik: probability of a mutation (line) to belong to clone k (column)
   *weights: proportion of mutations belonging to clone k
   *centers: cellularity of clone k in sample j
   *val: value of log-likelihood}
 \item{filtered.data}{input data without mutations removed (missing genotype or genotype not AB if filtered on AB)}
 \item{cluster}{Result of hard clustering: each mutation is attributed to a single clone}
}}
\usage{
QC_output
}
\description{
Clustering output generated by: 
QC_output<-One_step_clustering(SNV_list = Input_Example,contamination = c(0,0),
nclone_range = 2:5,maxit = 1,
save_plot = FALSE,ncores = 1,
epsilon = 0.01)
On October, 30th 2015
}
\keyword{datasets}

