% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumClone.R
\name{strcount}
\alias{strcount}
\title{String count}
\usage{
strcount(x, pattern = "", split = "")
}
\arguments{
\item{x}{The vector from which elements should be counted}

\item{pattern}{Pattern to be recognized. Default is ''}

\item{split}{Pattern used to split elements of the vector. Default is ''}
}
\description{
Counting the number of characters for each element of a vector
}
\keyword{Text}
\keyword{handling}

