% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantumClone.R
\name{Tidy_output}
\alias{Tidy_output}
\title{Tidying output from EM}
\usage{
Tidy_output(r, Genotype_provided, SNV_list)
}
\arguments{
\item{r}{output from Cluster_plot_from_cell}

\item{Genotype_provided}{If the FREEC_list is provided, then should be FALSE (default), otherwise TRUE}

\item{SNV_list}{A list of dataframes (one for each sample), with as columns : (for the first column of the first sample the name of the sample),}
}
\description{
Tidying input by Chr Start
}
