% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tree}
\alias{Tree}
\title{Example of output by Tree_generation}
\format{list of lists
\describe{
 \item{Top level}{List of all possibilities}
 \item{dataframe}{Table of hierarchical relations}
 \item{numeric}{probability of this tree}
}}
\usage{
Tree
}
\description{
Reconstruction of the tumor phylogenetic tree
Generated by :
 Tree<-Tree_generation(cbind(QuantumClone::QC_output$EM.output$centers[[1]],
 QuantumClone::QC_output$EM.output$centers[[2]]))
}
\keyword{datasets}

