% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tree_generation.R
\name{update_probs}
\alias{update_probs}
\title{Update proportions}
\usage{
update_probs(Proportions, can_split)
}
\arguments{
\item{Proportions}{numeric matrix}

\item{can_split}{logical vector}
}
\description{
Creates vector with probability to be sampled
}
