% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{zscore}
\alias{zscore}
\title{Z-score}
\usage{
zscore(Depth, Alt)
}
\arguments{
\item{Depth}{a numeric vector of depth of sequencing for each variant}

\item{Alt}{a numeric vector of the number of reads supporting each variant}
}
\value{
returns a square numeric matrix
}
\description{
Computes the z-score of mutations being from the same distribution
}

