% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EM.R
\name{BIC_criterion}
\alias{BIC_criterion}
\title{Bayesian Information Criterion}
\usage{
BIC_criterion(EM_out_list)
}
\arguments{
\item{EM_out_list}{list of outputs from EM.algo or FullEM}
}
\description{
Computes BIC from a list of outputs of EM algorithm, then returns the position with minimal BIC
}
\keyword{EM}
\keyword{clustering}
\keyword{number}

