% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{e.step}
\alias{e.step}
\title{Expectation step calculation}
\usage{
e.step(Schrod, centers, weights, adj.factor)
}
\arguments{
\item{Schrod}{A list of dataframes (one for each sample), generated by the Patient_schrodinger_cellularities() function.}

\item{centers}{Coordinates of the clones: a list of numeric vectors (1 per sample), with coordinates between 0 and 1.}

\item{weights}{Proportion of mutation in a clone}

\item{adj.factor}{Factor to compute the probability: makes transition between the cellularity of the clone and the frequency observed}
}
\description{
Expectation step calculation
}
\keyword{E-Step}
