% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descent.R
\name{eval.fik.m}
\alias{eval.fik.m}
\title{Eval probability for M step
Computes the log directly as log density is faster to compute}
\usage{
eval.fik.m(Schrod, centers, weights, adj.factor, log = TRUE)
}
\arguments{
\item{Schrod}{The shcrodinger list of matrices}

\item{centers}{centers of the clusters}

\item{weights}{weight of each cluster}

\item{adj.factor}{The adjusting factor, taking into account contamination, copy number, number of copies}

\item{log}{Should it compute the log distribution (TRUE) or probability (FALSE)
between two optimization steps. If NULL, will take 1/(median depth).}
}
\description{
Eval probability for M step
Computes the log directly as log density is faster to compute
}
