% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlashQC.R
\name{FlashQC}
\alias{FlashQC}
\title{Flash QuantumClone}
\usage{
FlashQC(Cells, conta, Nclus)
}
\arguments{
\item{Cells}{Input for QuantumClone with genotype required}

\item{conta}{vector with contamination fraction in each sample}

\item{Nclus}{vector with the number of clusters to test (alternatively only min and max values)}
}
\description{
Fast method to find clones without filtering for multiple states
}
\examples{
#1: Cluster data
FQC<-FlashQC(QuantumClone::Input_Example,conta = c(0,0),Nclus = 2:10)

}
\seealso{
QuantumClone
}

