% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_selection.R
\name{hard.clustering}
\alias{hard.clustering}
\title{Hard clustering based on EM output}
\usage{
hard.clustering(EM_out)
}
\arguments{
\item{EM_out}{Output from EM.algo or FullEM}
}
\description{
Attributes a mutation to its most likely clone based on the output of the EM algorithm
}
\keyword{EM}
\keyword{Hard}
\keyword{clustering}

