% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{manual_plot_trees}
\alias{manual_plot_trees}
\title{Plot tree}
\usage{
manual_plot_trees(connexion_list, d, cex = 0.8, p)
}
\arguments{
\item{connexion_list}{Data frame of the concatenation of the interaction matrix and the cellularity of each clone at different time points.}

\item{d}{Number of clusters found by QuantumClone}

\item{cex}{Coefficient of expansion for the texts in phylogenetic tree plots. Default is 0.8}

\item{p}{Probability of a tree}
}
\description{
Creates a visual output for the phylogeny created by Tree_generation()
}
\examples{
# Extract one tree out of the 3 available trees:
Example_tree<-QuantumClone::Tree[[1]]
manual_plot_trees(Example_tree[[1]], d= 4,p = Example_tree[[2]])
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}
