% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QuantumCat.R
\name{phylo_tree_generation}
\alias{phylo_tree_generation}
\title{Data generation}
\usage{
phylo_tree_generation(number_of_clones, number_of_samples)
}
\arguments{
\item{number_of_clones}{The wanted number of observable clones (meaning bearing at least 1 mutation)}

\item{number_of_samples}{The number of samples on which the data should be simulated}
}
\description{
WARNING: Tree_generation recreates a tree from data while phylo_tree_generation randomly creates a phylogeny
}
\keyword{Data}
\keyword{creation}
\keyword{phylogeny}

