% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_with_margins_densities}
\alias{plot_with_margins_densities}
\title{Plot with margin densities}
\usage{
plot_with_margins_densities(QClone_Output)
}
\arguments{
\item{QClone_Output}{Output from QuantumClone algorithm}
}
\description{
Adapted from http://stackoverflow.com/questions/11883844/inserting-a-table-under-the-legend-in-a-ggplot2-histogram
Uses gridExtra package
}
\examples{
require(ggplot2)
require(gridExtra)
message("Using preclustered data:")
QC_out<-QuantumClone::QC_output
plot_with_margins_densities(QC_out)
}
\keyword{Densities}
\keyword{Plot}

