% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phylogeny_tree.R
\name{Tree_generation}
\alias{Tree_generation}
\title{Phylogenetic tree}
\usage{
Tree_generation(Clone_cellularities, timepoints = NULL)
}
\arguments{
\item{Clone_cellularities}{A dataframe with cellularities (ranging from 0 to 1) of each clone (rows) in each sample (columns)}

\item{timepoints}{A numeric vector giving the spatial and/or temporal distribution of the samples}
}
\description{
Generates a list of possible trees based on the cellularity of each clone, and the spatial and temporal distribution of the samples.
Assumption is made the different clones are on different lines of the matrix
}
\keyword{Clonal}
\keyword{inference}
\keyword{phylogeny}
